requirements.txt
flask
manim
requests
python-dotenv
openai
langchain-core
graphviz